Amélioration de l'authentification en passant le token dans un cookie.²

+++ src/main/kotlin/perso/api/crypto/model/TransactionDto.kt
@@ -5,6 +5,6 @@ import java.time.LocalDateTime
 
 data class TransactionDto(
     val token: String,
-    val amount: BigDecimal,
+    val amount: Double,
     val dateTime: LocalDateTime
 )
